+++ quil-rs/src/instruction/mod.rs
@@ -459,11 +459,17 @@ impl Instruction {
                     FrameMatchCondition::AnyOfNames(frame_names),
                 ])
             }),
-            Instruction::Fence(Fence { qubits }) => Some(if qubits.is_empty() {
-                FrameMatchCondition::All
-            } else {
-                FrameMatchCondition::AnyOfQubits(Cow::Borrowed(qubits))
-            }),
+            Instruction::Fence(Fence { qubits }) => {
+                if include_blocked {
+                    Some(if qubits.is_empty() {
+                        FrameMatchCondition::All
+                    } else {
+                        FrameMatchCondition::AnyOfQubits(Cow::Borrowed(qubits))
+                    })
+                } else {
+                    None
+                }
+            }
             Instruction::Reset(Reset { qubit }) => {
                 let qubits = match qubit {
                     Some(qubit) => {
@@ -489,7 +495,7 @@ impl Instruction {
                 Some(FrameMatchCondition::Specific(frame))
             }
             Instruction::SwapPhases(SwapPhases { frame_1, frame_2 }) => {
-                Some(FrameMatchCondition::And(vec![
+                Some(FrameMatchCondition::Or(vec![
                     FrameMatchCondition::Specific(frame_1),
                     FrameMatchCondition::Specific(frame_2),
                 ]))
@@ -506,7 +512,6 @@ impl Instruction {
             | Instruction::Gate(_)
             | Instruction::GateDefinition(_)
             | Instruction::Halt
-            | Instruction::Wait
             | Instruction::Include(_)
             | Instruction::Jump(_)
             | Instruction::JumpUnless(_)
@@ -520,7 +525,8 @@ impl Instruction {
             | Instruction::Pragma(_)
             | Instruction::Store(_)
             | Instruction::UnaryLogic(_)
-            | Instruction::WaveformDefinition(_) => None,
+            | Instruction::WaveformDefinition(_)
+            | Instruction::Wait => None,
         }
     }
 
@@ -547,6 +553,53 @@ impl Instruction {
         let (_, instruction) =
             nom::combinator::all_consuming(parse_instruction)(&lexed).map_err(|e| e.to_string())?;
         Ok(instruction)
+    }
+
+    /// Per the Quil-T spec, whether this instruction's timing within the pulse
+    /// program must be precisely controlled so as to begin exactly on the end of
+    /// the latest preceding timed instruction
+    pub fn is_scheduled(&self) -> bool {
+        match self {
+            Instruction::Capture(_)
+            | Instruction::Delay(_)
+            | Instruction::Fence(_)
+            | Instruction::Pulse(_)
+            | Instruction::RawCapture(_)
+            | Instruction::SetFrequency(_)
+            | Instruction::SetPhase(_)
+            | Instruction::SetScale(_)
+            | Instruction::ShiftFrequency(_)
+            | Instruction::ShiftPhase(_)
+            | Instruction::SwapPhases(_)
+            | Instruction::Wait => true,
+            Instruction::Arithmetic(_)
+            | Instruction::BinaryLogic(_)
+            | Instruction::CalibrationDefinition(_)
+            | Instruction::CircuitDefinition(_)
+            | Instruction::Convert(_)
+            | Instruction::Comparison(_)
+            | Instruction::Declaration(_)
+            | Instruction::Exchange(_)
+            | Instruction::FrameDefinition(_)
+            | Instruction::Gate(_)
+            | Instruction::GateDefinition(_)
+            | Instruction::Halt
+            | Instruction::Include(_)
+            | Instruction::Jump(_)
+            | Instruction::JumpUnless(_)
+            | Instruction::JumpWhen(_)
+            | Instruction::Label(_)
+            | Instruction::Load(_)
+            | Instruction::MeasureCalibrationDefinition(_)
+            | Instruction::Measurement(_)
+            | Instruction::Move(_)
+            | Instruction::Nop
+            | Instruction::Pragma(_)
+            | Instruction::Reset(_)
+            | Instruction::Store(_)
+            | Instruction::UnaryLogic(_)
+            | Instruction::WaveformDefinition(_) => false,
+        }
     }
 }