+++ quil-rs/src/parser/expression.rs
@@ -39,12 +39,19 @@ enum Precedence {
 impl From<&Token> for Precedence {
     fn from(token: &Token) -> Self {
         match token {
-            Token::Operator(Operator::Plus) | Token::Operator(Operator::Minus) => Precedence::Sum,
-            Token::Operator(Operator::Star) | Token::Operator(Operator::Slash) => {
-                Precedence::Product
-            }
+            Token::Operator(operator) => Self::from(operator),
             // TODO: Is this used?
             Token::LParenthesis => Precedence::Call,
+            _ => Precedence::Lowest,
+        }
+    }
+}
+
+impl From<&Operator> for Precedence {
+    fn from(operator: &Operator) -> Self {
+        match operator {
+            Operator::Plus | Operator::Minus => Precedence::Sum,
+            Operator::Star | Operator::Slash => Precedence::Product,
             _ => Precedence::Lowest,
         }
     }
@@ -201,7 +208,7 @@ fn parse_infix(input: ParserInput, left: Expression) -> InternalParserResult<Exp
                 Operator::Slash => InfixOperator::Slash,
                 Operator::Star => InfixOperator::Star,
             };
-            let precedence = get_precedence(remainder);
+            let precedence = Precedence::from(token_operator);
             let (remainder, right) = parse(remainder, precedence)?;
             let infix_expression = Expression::Infix(InfixExpression {
                 left: Box::new(left),
@@ -227,12 +234,13 @@ fn parse_prefix(input: ParserInput) -> InternalParserResult<PrefixOperator> {
 
 #[cfg(test)]
 mod tests {
-    use crate::expression::{FunctionCallExpression, InfixExpression, PrefixExpression};
-    use crate::{expression::PrefixOperator, parser::lexer::lex};
-    use crate::{
-        expression::{Expression, ExpressionFunction, InfixOperator},
-        imag, real,
+    use crate::expression::{
+        Expression, ExpressionFunction, FunctionCallExpression, InfixExpression, InfixOperator,
+        PrefixExpression, PrefixOperator,
     };
+    use crate::instruction::MemoryReference;
+    use crate::parser::lexer::lex;
+    use crate::{imag, real};
 
     use nom_locate::LocatedSpan;
 
@@ -269,11 +277,13 @@ mod tests {
         let cases = vec![
             "pi",
             "sin(pi)",
-            "(1+(2*3))",
-            "((1+2)*3)",
+            "1+(2*3)",
+            "(1+2)*3",
             "%theta",
             "cis(%theta)",
-            "(%a+%b)",
+            "%a+%b",
+            "(pi/2)+(1*theta[0])",
+            "3--2",
         ];
 
         for case in cases {
@@ -352,6 +362,45 @@ mod tests {
             })),
             operator: InfixOperator::Star,
             right: Box::new(Expression::Variable("a".to_owned())),
+        })
+    );
+
+    test!(
+        infix_with_infix_operands_implicit_precedence,
+        parse_expression,
+        "pi/2 + 2*theta[0]",
+        Expression::Infix(InfixExpression {
+            left: Box::new(Expression::Infix(InfixExpression {
+                left: Box::new(Expression::PiConstant),
+                operator: InfixOperator::Slash,
+                right: Box::new(Expression::Number(real!(2f64))),
+            })),
+            operator: InfixOperator::Plus,
+            right: Box::new(Expression::Infix(InfixExpression {
+                left: Box::new(Expression::Number(real!(2f64))),
+                operator: InfixOperator::Star,
+                right: Box::new(Expression::Address(MemoryReference {
+                    name: "theta".to_string(),
+                    index: 0,
+                })),
+            })),
+        })
+    );
+
+    test!(
+        infix_minus_negative,
+        parse_expression,
+        "-3 - -2",
+        Expression::Infix(InfixExpression {
+            left: Box::new(Expression::Prefix(PrefixExpression {
+                operator: PrefixOperator::Minus,
+                expression: Box::new(Expression::Number(real!(3f64))),
+            })),
+            operator: InfixOperator::Minus,
+            right: Box::new(Expression::Prefix(PrefixExpression {
+                operator: PrefixOperator::Minus,
+                expression: Box::new(Expression::Number(real!(2.0)))
+            }))
         })
     );